@@ -1,6 +1,6 @@
 package godot.intellij.plugin.extension
 
-import java.util.Locale
+import java.util.*
 
 fun String.snakeToLowerCamelCase(): String {
     return "_[a-zA-Z]"
@@ -9,7 +9,7 @@ fun String.snakeToLowerCamelCase(): String {
             it
                 .value
                 .replace("_", "")
-                .toUpperCase()
+                .uppercase(Locale.US)
         }
 }
 
@@ -17,5 +17,7 @@ fun String.snakeToLowerCamelCase(): String {
 fun String.camelToSnakeCase(): String {
     return "(?<=[a-zA-Z0-9])[A-Z]".toRegex().replace(this) {
         "_${it.value}"
-    }.toLowerCase(Locale.ENGLISH)
+    }.lowercase(Locale.ENGLISH)
 }
+
+fun String.capitalize() = this.replaceFirstChar { if (it.isLowerCase()) it.titlecase(Locale.getDefault()) else it.toString() }
