@@ -3,8 +3,8 @@ package plugins.intellij
 data class BuildConfig(
     val sdk: String,
     val prefix: String,
-    val extraSource: String,
     val version: VersionRange,
     val ideVersionsForVerifierTask: List<String>,
-    val deps: List<String> = emptyList()
+    val deps: List<String> = emptyList(),
+    val extraSource: String? = null,
 )
