@@ -3,7 +3,7 @@ package godot.intellij.plugin.annotator.general
 import com.intellij.lang.annotation.AnnotationHolder
 import godot.intellij.plugin.GodotPluginBundle
 import godot.intellij.plugin.extension.registerProblem
-import org.jetbrains.kotlin.nj2k.postProcessing.type
+import godot.intellij.plugin.extension.type
 import org.jetbrains.kotlin.psi.KtProperty
 import org.jetbrains.kotlin.psi.KtTypeParameterListOwner
 import org.jetbrains.kotlin.types.typeUtil.containsTypeParameter
