@@ -3,12 +3,12 @@ package godot.intellij.plugin.annotator.reference
 import com.intellij.lang.annotation.AnnotationHolder
 import godot.intellij.plugin.GodotPluginBundle
 import godot.intellij.plugin.data.model.REGISTER_FUNCTION_ANNOTATION
-import godot.intellij.plugin.extension.getKotlinFqName
 import godot.intellij.plugin.extension.registerProblem
+import godot.intellij.plugin.extension.resolve
 import godot.intellij.plugin.quickfix.TargetFunctionNotRegisteredQuickFix
+import org.jetbrains.kotlin.idea.base.psi.kotlinFqName
 import org.jetbrains.kotlin.idea.util.findAnnotation
 import org.jetbrains.kotlin.name.FqName
-import org.jetbrains.kotlin.nj2k.postProcessing.resolve
 import org.jetbrains.kotlin.psi.KtCallExpression
 import org.jetbrains.kotlin.psi.KtCallableReferenceExpression
 import org.jetbrains.kotlin.psi.KtNameReferenceExpression
@@ -54,7 +54,7 @@ object RpcFunctionReferenceChecker {
                         ?.getArgumentExpression()
                         ?.getChildOfType<KtNameReferenceExpression>()
                         ?.resolve()
-                        ?.getKotlinFqName()
+                        ?.kotlinFqName
                         ?.asString() == "godot.MultiplayerAPI.RPCMode.DISABLED"
                 ) {
                     holder.registerProblem(
