@@ -5,6 +5,7 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
+import org.jetbrains.kotlin.idea.base.util.isGradleModule
 import org.jetbrains.kotlin.idea.core.util.toVirtualFile
 import org.jetbrains.kotlin.idea.util.projectStructure.getModule
 import org.jetbrains.kotlin.idea.util.projectStructure.module
@@ -14,7 +15,7 @@ import java.io.File
 typealias GodotRoot = String
 
 fun Module?.getGodotRoot(): GodotRoot? {
-    return if (this != null && !isDisposed && isGradleModule()) {
+    return if (this != null && !isDisposed && this.isGradleModule) {
         @Suppress("UnstableApiUsage")
         GradleUtil.findGradleModuleData(this)?.data?.let { moduleData ->
             File(moduleData.linkedExternalProjectPath).toVirtualFile()?.findChild("project.godot")?.path
