@@ -10,9 +10,9 @@ import godot.intellij.plugin.GodotPluginBundle
 import godot.intellij.plugin.data.cache.classname.RegisteredClassNameCacheProvider
 import godot.intellij.plugin.data.model.RegisteredClassDataContainer
 import godot.intellij.plugin.extension.getGodotRoot
-import org.jetbrains.kotlin.idea.core.util.getLineNumber
+import org.jetbrains.kotlin.idea.base.psi.getLineNumber
+import org.jetbrains.kotlin.idea.codeinsight.utils.findExistingEditor
 import org.jetbrains.kotlin.idea.core.util.toPsiFile
-import org.jetbrains.kotlin.idea.inspections.findExistingEditor
 import org.jetbrains.kotlin.psi.KtClass
 import org.jetbrains.kotlin.psi.KtElement
 import org.jetbrains.kotlin.psi.psiUtil.containingClass
@@ -62,7 +62,7 @@ class ClassAlreadyRegisteredQuickFix(private val registeredClassName: String) : 
                         .map { container -> container.fqName }
                         .toList()
                 )
-                .setTitle(GodotPluginBundle.message("quickFix.class.alreadyRegistered.popup.title"))
+            .setTitle(@Suppress("DialogTitleCapitalization") GodotPluginBundle.message("quickFix.class.alreadyRegistered.popup.title"))
                 .setItemChosenCallback { chosenFqName ->
                     val container = RegisteredClassNameCacheProvider
                         .provide(godotRoot)
